Fix errors with the server
