+++ package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.4.5",
     "env": "^0.0.2",
     "express": "^4.19.2",
     "gitignore": "^0.7.0"
