+++ public/index.html
@@ -4,14 +4,14 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Form Data Display</title>
-  <link rel="stylesheet" href="../client/style.css">
-  <script src="../client/client.js" async></script>
+  <link rel="stylesheet" href="/style.css">
+  <script src="/client.js" async></script>
 </head>
 <body>
 <header>
   <div class="header-style">
     <h1>Our ChatGPT Clone</h1>
-    <img src="../images/logo-black.png" class="img-style">  
+    <img src="images/logo-black.png" class="img-style">  
   </div>
 </header>
